state for offers
