+++ src/components/Offer/Offer.js
@@ -1,56 +1,21 @@
-import PropTypes from 'prop-types';
+import { useEffect, useState } from 'react';
+import { useParams } from 'react-router-dom';
+import { v4 as uuidv4 } from 'uuid';
 
-import { Button, Text, Input } from 'UI-KIT';
+import { apiGetOffersByCourse } from 'API/offers';
 
-import { BtnContainer } from 'COMPONENTS/CourseCard/styles/CourseCardStyled';
+import { Text, Input } from 'UI-KIT';
 
-import { Container, Description, Heading, OfferContainer, PreviewContainer, Title } from './styles/OfferStyle';
+import { OfferCard } from 'COMPONENTS/Offer/OfferCard';
 
-export const OfferCard = ({ offer, verticalAlign }) => (
-  <OfferContainer verticalAlign>
-    <Heading>
-      <Title>{offer.name}</Title>
-    </Heading>
-    {!verticalAlign && <Description>{offer.description}</Description>}
-    {verticalAlign && offer.benefits.map((el) => <li> {el} </li>)}
-    <BtnContainer>
-      <Button fullWidth>Купить</Button>
-    </BtnContainer>
-  </OfferContainer>
-);
-
+import { Container, PreviewContainer } from './styles/OfferStyle';
 export const Offer = () => {
-  const offers = [
-    {
-      description:
-        'Все предыдущие программы, которые мы писали, работали с текстовыми данными. Действительно, команда input() считывает строку текста.',
-      name: 'Язык Go Для Начинающих. Теория (Текст + Видео) + Разбор реального проекта + Доступ в закрытый чат.',
-      price: {
-        currency: 'usd',
-        value: 100,
-      },
-      benefits: ['benefit 1', 'benefit 2', 'benefit 3'],
-    },
-    {
-      description:
-        'Все предыдущие программы, которые мы писали, работали с текстовыми данными. Действительно, команда input() считывает строку текста.',
-      name: 'Язык Go Для Начинающих. Теория (Текст + Видео) + Разбор реального проекта.',
-      price: {
-        currency: 'usd',
-        value: 140,
-      },
-      benefits: ['benefit 1', 'benefit 2', 'benefit 3'],
-    },
-    {
-      description: 'Все предыдущие программы. Действительно, команда input() считывает строку текста.',
-      name: 'Язык Go Для Начинающих. Теория (Текст + Видео).',
-      price: {
-        currency: 'usd',
-        value: 10,
-      },
-      benefits: ['benefit 1', 'benefit 2', 'benefit 3'],
-    },
-  ];
+  const [offers, setOffers] = useState([]);
+  const { id } = useParams();
+
+  useEffect(() => {
+    apiGetOffersByCourse(id).then(({ data }) => setOffers(() => data));
+  }, []);
 
   return (
     <>
@@ -58,10 +23,10 @@ export const Offer = () => {
         <Text>Данный курс входит в следующие пакеты :</Text>
         <Input placeholder="Ввести промокод и получить скидку" />
       </PreviewContainer>
-      {offers.length - 1 ? (
+      {offers && offers.length - 1 ? (
         <Container>
           {offers.map((el) => (
-            <OfferCard key={el.name} offer={el} verticalAlign />
+            <OfferCard key={uuidv4()} offer={el} verticalAlign />
           ))}
         </Container>
       ) : (
@@ -69,14 +34,4 @@ export const Offer = () => {
       )}
     </>
   );
-};
-
-OfferCard.propTypes = {
-  offer: PropTypes.object,
-  verticalAlign: PropTypes.bool,
-};
-
-OfferCard.defaultProps = {
-  offer: null,
-  verticalAlign: false,
 };
